+++ tests/PHPUnit/Integration/Tracker/VisitTest.php
@@ -9,7 +9,6 @@
 namespace Piwik\Tests\Integration\Tracker;
 
 use Piwik\Cache;
-use Piwik\CacheId;
 use Piwik\Archive\ArchiveInvalidator;
 use Piwik\Date;
 use Piwik\Network\IPUtils;
@@ -17,12 +16,10 @@ use Piwik\Plugin\Manager;
 use Piwik\Plugins\SitesManager\API;
 use Piwik\Tests\Framework\Fixture;
 use Piwik\Tests\Framework\Mock\FakeAccess;
-use Piwik\Tracker\ActionPageview;
 use Piwik\Tracker\Request;
 use Piwik\Tracker\Visit;
 use Piwik\Tracker\VisitExcluded;
 use Piwik\Tests\Framework\TestCase\IntegrationTestCase;
-use Piwik\Tracker\Visitor;
 
 /**
  * @group Core
@@ -271,37 +268,6 @@ class VisitTest extends IntegrationTestCase
         }
     }
 
-    public function test_isVisitNew_ReturnsFalse_IfLastActionTimestampIsWithinVisitTimeLength_AndNoDimensionForcesVisit_AndVisitorKnown()
-    {
-        $this->setDimensionsWithOnNewVisit(array(false, false, false));
-
-        /** @var Visit $visit */
-        list($visit, $visitor, $action) = $this->makeVisitorAndAction(
-            $lastActionTime = '2012-01-02 08:08:34', $thisActionTime = '2012-01-02 08:12:45', $isVisitorKnown = true);
-
-        $result = $visit->isVisitNew($visitor, $action);
-
-        $this->assertFalse($result);
-    }
-
-    public function test_isVisitNew_ReturnsTrue_IfLastActionTimestampWasYesterday()
-    {
-        $this->setDimensionsWithOnNewVisit(array(false, false, false));
-
-        // test same day
-        /** @var Visit $visit */
-        list($visit, $visitor, $action) = $this->makeVisitorAndAction(
-            $lastActionTime = '2012-01-01 23:59:58', $thisActionTime = '2012-01-01 23:59:59', $isVisitorKnown = true);
-        $result = $visit->isVisitNew($visitor, $action);
-        $this->assertFalse($result);
-
-        // test different day
-        list($visit, $visitor, $action) = $this->makeVisitorAndAction(
-        $lastActionTime = '2012-01-01 23:59:58', $thisActionTime = '2012-01-02 00:00:01', $isVisitorKnown = true);
-        $result = $visit->isVisitNew($visitor, $action);
-        $this->assertTrue($result);
-    }
-
     public function test_markArchivedReportsAsInvalidIfArchiveAlreadyFinished_ShouldRemember_IfRequestWasDoneLongAgo()
     {
         $currentActionTime = '2012-01-02 08:12:45';
@@ -387,72 +353,6 @@ class VisitTest extends IntegrationTestCase
         $visit->handle();
 
         return array($visit, $request);
-    }
-
-    public function test_isVisitNew_ReturnsTrue_IfLastActionTimestampIsNotWithinVisitTimeLength_AndNoDimensionForcesVisit_AndVisitorNotKnown()
-    {
-        $this->setDimensionsWithOnNewVisit(array(false, false, false));
-
-        /** @var Visit $visit */
-        list($visit, $visitor, $action) = $this->makeVisitorAndAction($lastActionTime = '2012-01-02 08:08:34', $thisActionTime = '2012-01-02 09:12:45');
-
-        $result = $visit->isVisitNew($visitor, $action);
-
-        $this->assertTrue($result);
-    }
-
-    public function test_isVisitNew_ReturnsTrue_IfLastActionTimestampIsWithinVisitTimeLength_AndDimensionForcesVisit()
-    {
-        $this->setDimensionsWithOnNewVisit(array(false, false, true));
-
-        /** @var Visit $visit */
-        list($visit, $visitor, $action) = $this->makeVisitorAndAction($lastActionTime = '2012-01-02 08:08:34', $thisActionTime = '2012-01-02 08:12:45');
-
-        $result = $visit->isVisitNew($visitor, $action);
-
-        $this->assertTrue($result);
-    }
-
-    public function test_isVisitNew_ReturnsTrue_IfDimensionForcesVisit_AndVisitorKnown()
-    {
-        $this->setDimensionsWithOnNewVisit(array(false, false, true));
-
-        /** @var Visit $visit */
-        list($visit, $visitor, $action) = $this->makeVisitorAndAction($lastActionTime = '2012-01-02 08:08:34', $thisActionTime = '2012-01-02 08:12:45');
-
-        $result = $visit->isVisitNew($visitor, $action);
-
-        $this->assertTrue($result);
-    }
-
-    private function makeVisitorAndAction($lastActionTimestamp, $currentActionTime, $isVisitorKnown = false)
-    {
-        $idsite = API::getInstance()->addSite("name", "http://piwik.net/");
-
-        list($visit, $request) = $this->prepareVisitWithRequest(array('idsite' => $idsite), $currentActionTime);
-
-        $visitProperties = new Visit\VisitProperties();
-        $visitProperties->visitorInfo = array('visit_last_action_time' => Date::factory($lastActionTimestamp)->getTimestamp());
-
-        $visitor = new Visitor($request, 'configid', $visitProperties, $isVisitorKnown);
-
-        $action = new ActionPageview($request);
-
-        return array($visit, $visitor, $action);
-    }
-
-    private function setDimensionsWithOnNewVisit($dimensionOnNewVisitResults)
-    {
-        $dimensions = array();
-        foreach ($dimensionOnNewVisitResults as $onNewVisitResult) {
-            $dim = $this->getMock('Piwik\\Plugin\\Dimension', array('shouldForceNewVisit', 'getColumnName'));
-            $dim->expects($this->any())->method('shouldForceNewVisit')->will($this->returnValue($onNewVisitResult));
-            $dimensions[] = $dim;
-        }
-
-        $cache = Cache::getTransientCache();
-        $cache->save(CacheId::pluginAware('VisitDimensions'), $dimensions);
-        Visit::$dimensions = null;
     }
 
     public function provideContainerConfig()
